Remove engine folder and move all contents up a level
